Refactor: folder structure to seperate components and group common ones
