+++ my-app/src/Components/4 - RelatedComparison/Related/RelatedItems.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { ScrollMenu } from 'react-horizontal-scrolling-menu';
 import 'react-horizontal-scrolling-menu/dist/styles.css';
-import axiosAtelier from '../../axiosAtelier.js';
-import Card from './Card.jsx';
-import './hideScrollbar.css';
-import { LeftArrow, RightArrow } from './Arrow.jsx';
+import axiosAtelier from '../../../axiosAtelier.js';
+import Card from './RelatedCard.jsx';
+import '../Common/hideScrollbar.css';
+import { LeftArrow, RightArrow } from '../Common/Arrow.jsx';
 
 export default function App(ogProduct) {
   const [items, setItems] = React.useState([]);
